refactor(moderate): fix readRegister return type from long to uint8_t
- Return type now matches actual register size (8-bit)
- Eliminates unnecessary type promotion from uint8_t to long
- Also replaced magic number 0x10 with RREG constant
- API now correctly reflects that registers are 8-bit values

+++ src/ADS1256.h
@@ -110,7 +110,7 @@ static constexpr int8_t PIN_UNUSED = -1;
 	//ADS1256(int drate, int pga, int byteOrder, bool bufen);
 	
 	//Read a register
-	long readRegister(uint8_t registerAddress);
+	uint8_t readRegister(uint8_t registerAddress);
 	
 	//Write a register
 	void writeRegister(uint8_t registerAddress, uint8_t registerValueToWrite);
